move joint patterns to joint.py
